미완성

+++ 01_실습/src/Test4.java
@@ -2,6 +2,6 @@ public class Test4 {
     public static void main(String[] args) {
         int n = 1;
         int m = 2;
-        System.out.println(n / m);
+        System.out.println((double)n / m);
     }
 }
